流媒体安全问题

+++ services/media_engine/plugins/source/http_source/download/downloader.cpp
@@ -800,7 +800,7 @@ void Downloader::HandleRetOK()
     }
     int64_t remaining = 0;
     if (currentRequest_->endPos_ <= 0) {
-        remaining = currentRequest_->headerInfo_.fileContentLen > currentRequest_->startPos_?
+        remaining = static_cast<int64_t>(currentRequest_->headerInfo_.fileContentLen) > currentRequest_->startPos_?
             static_cast<int64_t>(currentRequest_->headerInfo_.fileContentLen) - currentRequest_->startPos_ : 0;
     } else {
         remaining = currentRequest_->endPos_ - currentRequest_->startPos_ + 1;
